菜单功能加入

+++ service/menu.go
@@ -28,7 +28,7 @@ func (menuService *MenuService) CreateMenu(req model.Menu) (model.Menu, error) {
 
 	var menu model.Menu
 	//验证菜单名称重复
-	if result := global.GaDb.Where(&model.Menu{Name: req.Name}).First(&menu); result.RowsAffected > 0 {
+	if result := global.GaDb.Where("name = ?", req.Meta.Name).First(&menu); result.RowsAffected > 0 {
 		return menu, business.New("该菜单名称已存在")
 	}
 	//创建菜单
@@ -37,9 +37,22 @@ func (menuService *MenuService) CreateMenu(req model.Menu) (model.Menu, error) {
 }
 
 func (menuService *MenuService) UpdateMenu(req model.Menu) error {
-	return global.GaDb.Updates(req).Error
+	updateMap := make(map[string]interface{})
+	updateMap["parent_id"] = req.ParentId
+	updateMap["route_path"] = req.RoutePath
+	updateMap["route_name"] = req.RouteName
+	updateMap["hidden"] = req.Hidden
+	updateMap["component"] = req.Component
+	updateMap["sort"] = req.Sort
+	updateMap["name"] = req.Name
+	updateMap["icon"] = req.Icon
+	return global.GaDb.Updates(updateMap).Error
 }
 
 func (menuService *MenuService) DeleteMenu(id int) error {
+	if result := global.GaDb.Where("parent_id = ?", id).First(&model.Menu{}); result.RowsAffected > 0 {
+		return business.New("此菜单存在子菜单不可删除")
+	}
+	//删除菜单及关联的角色
 	return global.GaDb.Delete(&model.Menu{}, id).Error
 }
